# Consistent fingerprint format and major speed improvement