This PR provides two things:

## Zero-padding of fingerprints

The other fingerprinting implementations (Java, Python, C, etc.) always zero-pad their hex strings, such as `85=06ece383,f87358c1`. This PR ensures that this is done in the PHP implementation as well (previously, the output would be `85=6ece383,f87358c1`.

## Speed improvement

By using the `hash('crc32c')` built-in in PHP 7.4, the performance can be greatly increased compared to using a user-provided function. See the benchmark below for the results of scanning the Linux kernel source, where the scanning time is reduced from 46 minutes to 10 minutes. The only downside is that this requires PHP 7.4 or later.

Before:
```
php -d memory_limit=8589934592 scanner.php /scratch/linux  2695.49s user 74.64s system 99% cpu 46:11.92 total
```
After:
```
php -d memory_limit=8589934592 scanner.php /scratch/linux  534.92s user 73.99s system 99% cpu 10:09.32 total
```

I have verified that the output is the same before and after my changes (well, apart from the zero-padded lines that are fixed, of course).